requirements.txt
boto3
langchain
langchain-community 
streamlit 
pillow 
faiss-cpu
unstructured